2-Django-Restframework-API/4-django-API-examples/requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
pipfile==0.0.2
pytz==2023.3
sqlparse==0.4.4
toml==0.10.2
